requirements.txt
boto
fabric
pyyaml
